requirements.txt
fastapi
pydantic
psycopg2
sqlalchemy
uvicorn
cryptography
redis
python-dotenv